content/canvas/test/test_size.attributes.removed.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.removed</title>
<!-- Testing: Removing content attributes reverts to default size -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="120" height="60"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

canvas.removeAttribute('width');
ok(canvas.width == 300, "canvas.width == 300");

SimpleTest.finish();

});
</script>
